==> 6. RS-forHomework.ipynb <==
# 将所有特征串联起来，构成RS_Train.csv
#RS_Test.csv
#为最后推荐系统做准备
from __future__ import division

import pickle
import numpy as np
import scipy.io as sio
import scipy.sparse as ss
import time
#相似度/距离
import scipy.spatial.distance as ssd

from numpy.random import random  
from collections import defaultdict

class RecommonderSystem:
  def __init__(self):
    # 读入数据做初始化
    
    #用户和活动新的索引
    self.userIndex = pickle.load(open("PE_userIndex.pkl", 'rb'))
    self.eventIndex = pickle.load(open("PE_eventIndex.pkl", 'rb'))
    self.n_users = len(self.userIndex)
    self.n_items = len(self.eventIndex)
    
    #用户-活动关系矩阵R
    #在train_SVD会重新从文件中读取,二者要求的格式不同，来不及统一了
    self.userEventScores = sio.mmread("PE_userEventScores").todense()
    
    #倒排表
    ##每个用户参加的事件
    self.itemsForUser = pickle.load(open("PE_eventsForUser.pkl", 'rb'))
    ##事件参加的用户
    self.usersForItem = pickle.load(open("PE_usersForEvent.pkl", 'rb'))
    
    #基于模型的协同过滤参数初始化,训练
    self.init_SVD()
    self.train_SVD(trainfile = "train.csv")
    
    #根据用户属性计算出的用户之间的相似度
    self.userSimMatrix = sio.mmread("US_userSimMatrix").todense()
    
    #根据活动属性计算出的活动之间的相似度
    self.eventPropSim = sio.mmread("EV_eventPropSim").todense()  #非词频特征，采用Person相关系数作为相似度
    self.eventContSim = sio.mmread("EV_eventContSim").todense()  #对词频特征，采用余弦相似度，也可以用直方图交/Jacard相似度
    
    #根据用户打分计算出的用户之间的相似度
    self.userCFSim = sio.mmread("US_userCFSim").todense()
    
    #根据用户打分计算出的活动之间的相似度
    self.eventCFSim = sio.mmread("EV_eventCFSim").todense()
    
    #每个用户的朋友的数目
    self.numFriends = sio.mmread("UF_numFriends")
    #用户的每个朋友参加活动的分数对该用户的影响
    self.userFriends = sio.mmread("UF_userFriends").todense()
    
    #活动本身的热度
    self.eventPopularity = sio.mmread("EA_eventPopularity").todense()

  def init_SVD(self, K=20):
    #初始化模型参数（for 基于模型的协同过滤SVD_CF）
    self.K = K  
    
    #init parameters
    #bias
    self.bi = np.zeros(self.n_items)  
    self.bu = np.zeros(self.n_users)  
    
    #the small matrix
    self.P = random((self.n_users,self.K))/10*(np.sqrt(self.K))
    self.Q = random((self.K, self.n_items))/10*(np.sqrt(self.K))  
                  
          
  def train_SVD(self,trainfile = 'train.csv', steps=10,gamma=0.04,Lambda=0.15):
    #训练SVD模型（for 基于模型的协同过滤SVD_CF）
    #gamma：为学习率
    #Lambda：正则参数
    
    #偷懒了，为了和原来的代码的输入接口一样，直接从训练文件中去读取数据
    print ("SVD Train...")
    ftrain = open(trainfile, 'r')
    ftrain.readline() # skip header 跳过第一行（表头）
    self.mu = 0.0
    n_records = 0
    uids = []  #每条记录的用户索引
    i_ids = [] #每条记录的item索引
    #用户-Item关系矩阵R（内容同userEventScores相同），临时变量，训练完了R不再需要
    R = np.zeros((self.n_users, self.n_items))  # n_users用户数量  n_items活动数量
    
    for line in ftrain:
        cols = line.strip().split(",")
        u = self.userIndex[cols[0]]  #用户Index
        i = self.eventIndex[cols[1]] #活动Index
        
        uids.append(u)  #记录用户索引
        i_ids.append(i)  #记录item（活动）索引
        
        R[u,i] = int(cols[4]) - int(cols[5])  #interested - not_interested
        self.mu += R[u,i]
        n_records += 1
    
    ftrain.close()
    self.mu /= n_records
    # ============================ 补充代码 =================================
    # ======================== SVD模型训练过程 ==============================
    # 去除用户、活动索引数组中重复元素
    uids = list(set(uids))
    i_ids = list(set(i_ids))
    
#     print('len(uid)', len(uids))
#     print('len(i_id)', len(i_ids))
    for step in range(steps):  
        print ('the ',step,'-th  step is running') 
        time_start=time.time()
        rmse_sum=0.0 

        #将训练样本打散顺序
        randomUserIndex = np.random.permutation(len(uids))
        randomItemIndex = np.random.permutation(len(i_ids))
        n_iter = 0
        for i in range(len(uids)):
            for j in range(len(i_ids)):
                uid = uids[randomUserIndex[i]]
                i_id = i_ids[randomItemIndex[j]]
                rat = R[uid][i_id]

                #预测残差
                eui=rat-self.pred_SVD(uid,i_id)
                if eui == 0:
                    continue

                #残差平方和
                rmse_sum+=eui**2  

                #随机梯度下降，更新
                self.bu[uid]+=gamma*(eui-Lambda*self.bu[uid])  
                self.bi[i_id]+=gamma*(eui-Lambda*self.bi[i_id]) 
               
                temp=self.Q[:,i_id]  
                self.Q[:,i_id]+=gamma*(eui*self.P[uid,:]-Lambda*self.Q[:,i_id])  
                self.P[uid,:]+=gamma*(eui*temp-Lambda*self.P[uid,:])
                
                n_iter += 1
                
#                 if n_iter == 10:
#                     break
#             if n_iter == 10:
#                 break
                
        #学习率递减
        gamma=gamma*0.93  
        time_end=time.time()
#         print('rmse_sum',rmse_sum)
        print('n_iter', n_iter)
        print ("the rmse of this step on train data is ",np.sqrt(rmse_sum/n_iter),
               '  this step cost ',(time_end-time_start) // 60,'m', (time_end-time_start) % 60, 's') 
        # ===================================================================
    
    print ("SVD trained")
    
  def pred_SVD(self, uid, i_id):
    #根据当前参数，预测用户uid对Item（i_id）的打分   
    #uid、i_id分别为user和item的索引
    ans=self.mu + self.bi[i_id] + self.bu[uid] + np.dot(self.P[uid,:],self.Q[:,i_id])  
        
    #将打分范围控制在0-1之间
    if ans>1:  
        return 1  
    elif ans<0:  
        return 0
    return ans  

  def sim_cal_UserCF(self, uid1, uid2 ):
    #请补充基于用户的协同过滤中的两个用户uid1和uid2之间的相似度（根据两个用户对item打分的相似度）
    similarity = 0.0
    # ================================= 补充代码 =====================================
    # ============================== 两个用户间相似度 =================================
    #检查用户ID是否有效
    if uid1 not in self.userIndex or uid2 not in self.userIndex:
        return 0.0
    #相同id用户相似度为1
    if uid1 == uid2:
        return 1.0
    
    i = self.userIndex[uid1]
    j = self.userIndex[uid2]
    
    similarity = ssd.correlation(self.userEventScores[i, :],self.userEventScores[j, :])
    # =================================================================================
    return similarity  

  def userCFReco(self, userId, eventId):
    """
    根据User-based协同过滤，得到event的推荐度
    基本的伪代码思路如下：
    for item i
      for every other user v that has a preference for i
        compute similarity s between u and v
        incorporate v's preference for i weighted by s into running aversge
    return top items ranked by weighted average
    """
    #请补充完整代码
    # ================================= 补充代码 =====================================
    # ================= 根据User-based协同过滤，得到event的推荐度 ====================
    i = self.userIndex[userId]
    j = self.eventIndex[eventId]
    userIdlist = list(self.userIndex.keys())

    vs = self.userEventScores[:, j]
    sims = ss.dok_matrix((1, self.n_users))
    
    # 挨个计算用户与用户间的相似度，巨慢！！！
#     for uindex in range(self.n_users):
#         sims[0,uindex] = self.sim_cal_UserCF(userId, userIdlist[uindex])
    #使用先前根据用户打分计算的用户相似度矩阵获取当前用户与其他用户间相似度 （速度快）
    sims = self.userCFSim[i, :]
    
    prod = sims * vs

    try:
      return prod[0, 0] - self.userEventScores[i, j]
    except IndexError:
      return 0    
    # =================================================================================

  def sim_cal_ItemCF(self, i_id1, i_id2):
    #计算Item i_id1和i_id2之间的相似性
    #请补充完整代码
    similarity = 0.0
    # ================================= 补充代码 =====================================
    # ============================== 两个item间相似度 =================================
    
    sim_cal_ItemCF_start = time.time()
    
    #检查活动ID是否有效
    if i_id1 not in self.eventIndex or i_id2 not in self.eventIndex:
        return 0.0
    #相同id活动相似度为1
    if i_id1 == i_id2:
        return 1.0
    
    i = self.eventIndex[i_id1]
    j = self.eventIndex[i_id2]
    similarity = ssd.correlation(self.userEventScores[:,i],self.userEventScores[:,j])
    
    sim_cal_ItemCF_end = time.time()
    print('sim_cal_ItemCF cast', sim_cal_ItemCF_end-sim_cal_ItemCF_start)
    
    return similarity     
    # =================================================================================      
  def eventCFReco(self, userId, eventId):    
    """
    根据基于物品的协同过滤，得到Event的推荐度
    基本的伪代码思路如下：
    for item i 
        for every item j tht u has a preference for
            compute similarity s between i and j
            add u's preference for j weighted by s to a running average
    return top items, ranked by weighted average
    """
    #请补充完整代码
    # ================================= 补充代码 =====================================
    # ================== 根据基于物品的协同过滤，得到Event的推荐度 ===================
    i = self.userIndex[userId]
    j = self.eventIndex[eventId]
    eventIdlist = list(self.eventIndex.keys())
    
    js = self.userEventScores[i, :]
    sims = ss.dok_matrix((self.n_items, 1))
    
    # 挨个计算活动与活动间的相似度，巨慢！！！
#     for eindex in range(self.n_items):
#         sims[eindex, 0] = self.sim_cal_ItemCF(eventId, eventIdlist[eindex])
    #使用根据用户打分计算的用户相似度矩阵，获取当前活动与其他活动间相似度 （速度快）
    sims = self.eventCFSim[:, j]
    
    prod = js * sims    
    ans = 0.0
    try:
      ans = prod[0, 0] - self.userEventScores[i, j]
    except IndexError:
      pass
    return ans     
    # =================================================================================
    
  def svdCFReco(self, userId, eventId):
    #基于模型的协同过滤, SVD++/LFM
    u = self.userIndex[userId]
    i = self.eventIndex[eventId]

    return self.pred_SVD(u,i)

  def userReco(self, userId, eventId):
    """
    类似基于User-based协同过滤，只是用户之间的相似度由用户本身的属性得到，计算event的推荐度
    基本的伪代码思路如下：
    for item i
      for every other user v that has a preference for i
        compute similarity s between u and v
        incorporate v's preference for i weighted by s into running aversge
    return top items ranked by weighted average
    """
    i = self.userIndex[userId]
    j = self.eventIndex[eventId]

    vs = self.userEventScores[:, j]
    sims = self.userSimMatrix[i, :]

    prod = sims * vs

    try:
      return prod[0, 0] - self.userEventScores[i, j]
    except IndexError:
      return 0

  def eventReco(self, userId, eventId):
    """
    类似基于Item-based协同过滤，只是item之间的相似度由item本身的属性得到，计算Event的推荐度
    基本的伪代码思路如下：
    for item i 
      for every item j that u has a preference for
        compute similarity s between i and j
        add u's preference for j weighted by s to a running average
    return top items, ranked by weighted average
    """
    i = self.userIndex[userId]
    j = self.eventIndex[eventId]
    js = self.userEventScores[i, :]
    psim = self.eventPropSim[:, j]
    csim = self.eventContSim[:, j]
    pprod = js * psim
    cprod = js * csim
    
    pscore = 0
    cscore = 0
    try:
      pscore = pprod[0, 0] - self.userEventScores[i, j]
    except IndexError:
      pass
    try:
      cscore = cprod[0, 0] - self.userEventScores[i, j]
    except IndexError:
      pass
    return pscore, cscore

  def userPop(self, userId):
    """
    基于用户的朋友个数来推断用户的社交程度
    主要的考量是如果用户的朋友非常多，可能会更倾向于参加各种社交活动
    """
    if userId in self.userIndex:
      i = self.userIndex[userId]
      try:
        return self.numFriends[0, i]
      except IndexError:
        return 0
    else:
      return 0

  def friendInfluence(self, userId):
    """
    朋友对用户的影响
    主要考虑用户所有的朋友中，有多少是非常喜欢参加各种社交活动/event的
    用户的朋友圈如果都积极参与各种event，可能会对当前用户有一定的影响
    """
    nusers = np.shape(self.userFriends)[1]
    i = self.userIndex[userId]
    return (self.userFriends[i, :].sum(axis=0) / nusers)[0,0]

  def eventPop(self, eventId):
    """
    本活动本身的热度
    主要是通过参与的人数来界定的
    """
    i = self.eventIndex[eventId]
    return self.eventPopularity[i, 0]



def generateRSData(RS, train=True, header=True):
    """
    把前面user-based协同过滤 和 item-based协同过滤，以及各种热度和影响度作为特征组合在一起
    生成新的训练数据，用于分类器分类使用
    """
    fn = "train.csv" if train else "test.csv"
    fin = open(fn, 'r')
    fout = open("RS_" + fn, 'w')
    
    #忽略第一行（列名字）
    fin.readline().strip().split(",")
    
    # write output header
    if header:
      ocolnames = ["invited", "userCF_reco", "evtCF_reco","svdCF_reco","user_reco", "evt_p_reco",
        "evt_c_reco", "user_pop", "frnd_infl", "evt_pop"]
      if train:
        ocolnames.append("interested")
        ocolnames.append("not_interested")
      fout.write(",".join(ocolnames) + "\n")
    
    ln = 0
    for line in fin:
        ln += 1
        if ln%500 == 0:
          print ("%s:%d (userId, eventId)=(%s, %s)" % (fn, ln, userId, eventId))
          #break;
#         lntime_start = time.time()

        cols = line.strip().split(",")
        userId = cols[0]
        eventId = cols[1]
        invited = cols[2]
        
#         lntime_part = time.time()
        userCF_reco = RS.userCFReco(userId, eventId)
#         userCF_recotime = time.time()
#         print('this ln(',ln,') userCF_recotime ',(userCF_recotime-lntime_part) // 60,'m', (userCF_recotime-lntime_part) % 60, 's')
        
        itemCF_reco = RS.eventCFReco(userId, eventId)
#         itemCF_recotime = time.time()
#         print('this ln(',ln,') itemCF_recotime ',(itemCF_recotime-userCF_recotime) // 60,'m', (itemCF_recotime-userCF_recotime) % 60, 's')
        
        svdCF_reco = RS.svdCFReco(userId, eventId)
#         svdCF_recotime = time.time()
#         print('this ln(',ln,') svdCF_recotime ',(svdCF_recotime-itemCF_recotime) // 60,'m', (svdCF_recotime-itemCF_recotime) % 60, 's')

        user_reco = RS.userReco(userId, eventId)
        evt_p_reco, evt_c_reco = RS.eventReco(userId, eventId)
        user_pop = RS.userPop(userId)

        frnd_infl = RS.friendInfluence(userId)
        evt_pop = RS.eventPop(eventId)
        ocols = [invited, userCF_reco, itemCF_reco, svdCF_reco,user_reco, evt_p_reco,
        evt_c_reco, user_pop, frnd_infl, evt_pop]

        if train:
            ocols.append(cols[4]) # interested
            ocols.append(cols[5]) # not_interested
            fout.write(",".join(map(lambda x: str(x), ocols)) + "\n")
            
#         lntime_end = time.time()
#         print('this ln(',ln,') part cost ',(lntime_end-lntime_part) // 60,'m', (lntime_end-lntime_part) % 60, 's')
#         print('this ln(',ln,') cost ',(lntime_end-lntime_start) // 60,'m', (lntime_end-lntime_start) % 60, 's')
    
    fin.close()
    fout.close()


RS = RecommonderSystem()

print ("生成训练数据...\n")
generateRSData(RS,train=True,  header=True)

print ("生成预测数据...\n")
generateRSData(RS, train=False, header=True)

时间、地点等特征都没有处理了，可以考虑用户看到event的时间与event开始时间的差、用户地点和event地点的差异。。。
